Clone War's. A simple tool to find and remove idendical files.
Clone War's is a bash- shell script: 'compare.sh'
It checks the dir it resides in and all of it's sub dir's for binary idendical files (unix's 'diff'- command).
A protocol- file is created and all idendical files are moved to a directory created by the script ('dupDir').

NO FILES ARE DELETED! You have to do that manually by checking 'dupDir' and the protocol file.

Berthold Fritz 2018